Make the dataset loading common
