+++ cal_ratio_trainer/config.py
@@ -1,8 +1,6 @@
 import io
-import logging
 from pathlib import Path
 from typing import List, Optional
-import fsspec
 from pydantic import AnyUrl, BaseModel, Field
 import yaml
 
@@ -59,22 +57,6 @@ class TrainingConfig(BaseModel):
     main_training_file: Optional[AnyUrl] = None
     cr_training_file: Optional[AnyUrl] = None
 
-    data_cache: Optional[Path] = None
-
-    @property
-    def main_file(self) -> Path:
-        assert (
-            self.data_cache is not None
-        ), "Must have a valid data_cache in configuration parameters"
-        return make_local(str(self.main_training_file), self.data_cache)
-
-    @property
-    def cr_file(self) -> Path:
-        assert (
-            self.data_cache is not None
-        ), "Must have a valid data_cache in configuration parameters"
-        return make_local(str(self.cr_training_file), self.data_cache)
-
     def __str__(self) -> str:
         string_out = io.StringIO()
         for k, v in self.dict().items():
@@ -101,38 +83,3 @@ def load_config(p: Optional[Path] = None) -> TrainingConfig:
                 setattr(r, k, v)
 
     return r
-
-
-def make_local(file_path: str, cache: Path) -> Path:
-    """Uses the `fsspec` library to copy a non-local file locally in the `cache`.
-    If the `file_path` is already a local file, then it isn't copied locally.
-
-    Args:
-        file_path (str): The URI of the file we want to be local.
-
-    Returns:
-        Path: Path on the local system to the data.
-    """
-    if file_path.startswith("file://"):
-        return Path(file_path[7:])
-    else:
-        local_path = cache / Path(file_path).name
-        if local_path.exists():
-            return local_path
-
-        # Ok - copy block by block.
-        logging.warning(f"Copying file {file_path} locally to {cache}")
-        local_path.parent.mkdir(parents=True, exist_ok=True)
-
-        tmp_file_path = local_path.with_suffix(".tmp")
-        with open(tmp_file_path, "wb") as f_out:
-            with fsspec.open(file_path, "rb") as f_in:
-                # Read `f_in` in chunks of 1 MB and write them to `f_out`
-                while True:
-                    data = f_in.read(50 * 1024**2)  # type: ignore
-                    if not data:
-                        break
-                    f_out.write(data)
-        tmp_file_path.rename(local_path)
-        logging.warning(f"Done copying file {file_path}")
-        return local_path
